+++ app/src/main/java/com/github/libretube/MyApp.kt
@@ -5,8 +5,6 @@ import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.os.Build
 
-const val NOTIFICATION_ID = 1
-
 class MyApp : Application() {
     override fun onCreate() {
         super.onCreate()
@@ -15,19 +13,37 @@ class MyApp : Application() {
     }
 
     /**
-     * Initializes the required [NotificationChannel] for the app.
+     * Initializes the required [NotificationChannel]s for the app.
      */
     private fun initializeNotificationChannels() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            // Create the NotificationChannel
-            val name = "Background Mode"
-            val descriptionText = "Shows a notification with buttons to control the audio player"
-            val importance = NotificationManager.IMPORTANCE_LOW
-            val mChannel = NotificationChannel("background_mode", name, importance)
-            mChannel.description = descriptionText
+        createNotificationChannel(
+            "download_service",
+            "Download Service",
+            "DownloadService",
+            NotificationManager.IMPORTANCE_NONE
+        )
+        createNotificationChannel(
+            "background_mode",
+            "Background Mode",
+            "Shows a notification with buttons to control the audio player",
+            NotificationManager.IMPORTANCE_LOW
+        )
+    }
 
+    private fun createNotificationChannel(
+        id: String,
+        name: String,
+        descriptionText: String,
+        importance: Int
+    ) {
+        // Create the NotificationChannel, but only on API 26+ because
+        // the NotificationChannel class is new and not in the support library
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            val channel = NotificationChannel(id, name, importance)
+            channel.description = descriptionText
+            // Register the channel in the system
             val notificationManager = getSystemService(NOTIFICATION_SERVICE) as NotificationManager
-            notificationManager.createNotificationChannel(mChannel)
+            notificationManager.createNotificationChannel(channel)
         }
     }
